Some DHCPv4 fixes
* Increment cursor after adding static routes
* Fix log message for unhandled DHCPv4 packets

+++ agent-ovs/src/PacketInHandler.cpp
@@ -506,7 +506,7 @@ static void handleDHCPv4PktIn(shared_ptr<const Endpoint>& ep,
         break;
     default:
         LOG(DEBUG) << "Ignoring unexpected DHCP message of type "
-                   << message_type << " from " << srcMac;
+                   << (uint16_t)message_type << " from " << srcMac;
         return;
     }
